test: create test to assert user revoked token
